- Ajustes da scene StatementsList - Passando dados da tela de Login para tela de StatementsList

+++ Bank_App/Bank_App/Bank_App/Scenes/Login/LoginModels.swift
@@ -41,7 +41,7 @@ struct UserAccount: Codable {
   let name: String?
   let bankAccount: String?
   let agency: String?
-  let balance: Double?
+  let balance: Float?
 }
 
 struct Error: Codable {
